Manually updated (as much as possible) orphaned/unmaintained languages
For languages whose maintainers have not responded with regards to the
changes in EasyBCD 2.3 (ca, es, he, hu, ko, pl, tr, zh-CHS), they have
been manually updated as follows:

* Where a near-definitive translation could be located online for keys
  (mostly for minor changes or single-word translations), the update
  was made and the translation for that key was marked as synced and
  correct.
* Where a non-complete translation was made that suffices to inform the
  user of the correct intent behind a feature could be made but that
  does not fully convey the meaning of the original English text, the
  change was made but the key has remained marked as out-of-date for
  future reference.
* Where a translation could not be made without help of a native
  speaker and the existing translation is outright incorrect or
  invalid, the key was removed and the English text will substitute in
  its place until such a time that the correct translation can be
  made/obtained.

For the Hebrew (he) translation: this is a very incomplete translation
and is at risk of being removed in the future. It has been updated
according to the instructions outlined above, but as much of the
non-UI text remains untranslated, it is in need of a new maintainer and
serious attention.